requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0.0
paho-mqtt
requests
